epicevents/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
sentry-sdk==1.10.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
